PR: move cliplugins in cmd to pkg
