+++ pkg/cliplugins/workload/plugin/edge-syncer.yaml
